Bring back lib files
