+++ lib/modules/updates/widgets/status_icon.dart
