+++ lib/shared/contents.dart
