+++ lib/shared/widgets/card_preview.dart
